idk bro, i'm just pushing shit that works :)

+++ dataBaseManager/src/main/java/edu/upvictoria/fpoo/utils/CommandInterpreter.java
@@ -6,10 +6,7 @@ import edu.upvictoria.fpoo.exceptions.SQLSyntaxException;
 import javax.naming.NoPermissionException;
 import java.io.*;
 import java.nio.file.NotDirectoryException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -55,7 +52,14 @@ public class CommandInterpreter {
      */
     public Map<String, String> parseSQLQuery(String sqlQuery) {
         Map<String, String> resultMap = new HashMap<>();
-        Pattern pattern = Pattern.compile("(?i)(SELECT|FROM|WHERE|ORDER BY|CREATE TABLE|INSERT|UPDATE|DELETE|DROP TABLE|SHOW TABLES)\\s+(.*?)(?=\\s+(?i)(?:SELECT|FROM|WHERE|ORDER BY|CREATE TABLE|INSERT|UPDATE|DELETE|DROP TABLE|SHOW TABLES)|;|$)");
+        if (sqlQuery.trim().toLowerCase().startsWith("use")) {
+            String key = "USE";
+            String value = sqlQuery.trim().split(" ")[1].replace(";", "");
+            resultMap.put(key, value);
+            return resultMap;
+        }
+
+        Pattern pattern = Pattern.compile("(?i)(SELECT|FROM|WHERE|ORDER BY|CREATE TABLE|INSERT INTO|UPDATE|DELETE|DROP TABLE|SHOW TABLES)\\s+(.*?)(?=\\s+(?i)(?:SELECT|FROM|WHERE|ORDER BY|CREATE TABLE|INSERT INTO|UPDATE|DELETE|DROP TABLE|SHOW TABLES)|;|$)");
         Matcher matcher = pattern.matcher(sqlQuery);
 
         while (matcher.find()) {
@@ -82,12 +86,40 @@ public class CommandInterpreter {
             }
 
             if (command.equals("DROP TABLE")) {
-                dropTable(value);
+                try {
+                    dropTable(value);
+                } catch (SQLSyntaxException e) {
+                    System.err.println("SQLSyntaxException: " + e.getMessage());
+                }
                 return;
             }
 
+            if (command.equals("INSERT INTO")) {
+                try {
+                    insert(value);
+                } catch (SQLSyntaxException e) {
+                    System.err.println("SQLSyntaxException: " + e.getMessage());
+                }
+                return;
+            }
+
+            if (command.equalsIgnoreCase("USE")) {
+                try {
+                    use(value);
+                } catch (NotDirectoryException | FileNotFoundException | NoPermissionException | NotADBException e) {
+                    System.err.println(e.getMessage());
+                }
+                return;
+            }
             throw new SQLSyntaxException("Not a valid SQL syntax");
         }
+    }
+
+    public void use(String folderPath) throws NotADBException, NotDirectoryException, FileNotFoundException, NoPermissionException {
+        if (!folderPath.endsWith("_db"))
+            throw new NotADBException("The folder is not a known database");
+        this.folder = new IOUtils().openFolder(folderPath);
+        System.out.println(ConsoleColors.ANSI_BLUE + "Using " + folderPath + ConsoleColors.ANSI_RESET);
     }
 
     public void createTable(String tableParams) throws SQLSyntaxException {
@@ -97,7 +129,7 @@ public class CommandInterpreter {
         if (!matcher.find())
             throw new SQLSyntaxException("Not a valid sql syntax");
 
-        String tableName = matcher.group(1);
+        String tableName = matcher.group(1).trim();
         String columnsText = matcher.group(2);
 
         String[] columns = columnsText.toLowerCase().split(",");
@@ -124,7 +156,7 @@ public class CommandInterpreter {
         }
 
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(csvPath))) {
-            int i =  0;
+            int i = 0;
             for (String column : columnsList) {
                 writer.write(column);
 
@@ -132,16 +164,26 @@ public class CommandInterpreter {
                     writer.write(",");
                 i++;
             }
+            writer.newLine();
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            System.err.println("There was an error creating the table");
         }
+        System.out.println(ConsoleColors.ANSI_CYAN + "Table " + tableName + " successfully created" + ConsoleColors.ANSI_RESET);
     }
 
-    public void dropTable (String tableName) throws SQLSyntaxException {
+    public void dropTable(String tableName) throws SQLSyntaxException {
         if (tableName.split(" ").length > 1)
             throw new SQLSyntaxException("Not a valid table name");
 
         String csvPath = folder + File.separator + tableName + ".csv";
+
+        System.out.println("Are you sure you want to delete this database?");
+        System.out.println("1. Yes");
+        System.out.println("2. No");
+
+        if (BrScanner.readInt() == 2)
+            return;
+
 
         try {
             FileUtils.deleteFile(csvPath);
@@ -150,6 +192,98 @@ public class CommandInterpreter {
         }
 
         System.out.println("Table dropped successfully");
+    }
+
+    public void insert(String commandParams) throws SQLSyntaxException {
+        commandParams = commandParams.toLowerCase();
+
+        Pattern pattern = Pattern.compile("(?i)\\b(\\w+)\\b\\s*\\((.*?)\\)\\s*values\\s*\\((.*?)\\)");
+        Matcher matcher = pattern.matcher(commandParams);
+
+        if (!matcher.find())
+            throw new SQLSyntaxException("Not a valid sql syntax");
+
+        String tableName = matcher.group(1);
+        String columnNames = matcher.group(2);
+        String values = matcher.group(3);
+
+        if (columnNames.split(",").length != values.split(",").length)
+            throw new SQLSyntaxException("The number of columns and values does not match");
+
+
+        String tablePath = folder + File.separator + tableName + ".csv";
+        List<String> columnsList = getTableColumns(tablePath);
+        List<String> columns = new ArrayList<>(List.of(columnNames.split(",")));
+        columns.replaceAll(String::trim);
+
+        List<String> valuesList = new ArrayList<>(List.of(values.split(",")));
+        valuesList.replaceAll(String::trim);
+
+        if (!ListUtils.listEqualsIgnoreOrder(columns, columnsList))
+            throw new SQLSyntaxException("Columns do not match");
+
+        List<String> rightValues = new ArrayList<>();
+
+        for (String column : columnsList) {
+            for (String column2 : columns) {
+                if (column.equals(column2)) {
+                    int index = columns.indexOf(column2);
+                    rightValues.add(valuesList.get(index));
+                }
+            }
+        }
+
+        StringBuilder lineContent = new StringBuilder();
+
+        int i = 0;
+        for (String value : rightValues) {
+            lineContent.append(value);
+            if (i < rightValues.size() - 1)
+                lineContent.append(",");
+            i++;
+        }
+
+        String csvPath = folder + File.separator + tableName + ".csv";
+        FileUtils.addToFile(csvPath, String.valueOf(lineContent));
+    }
+
+    public List<String> getTableColumns(String tablePath) throws SQLSyntaxException {
+        try (BufferedReader br = new BufferedReader(new FileReader(tablePath))) {
+            String line = br.readLine();
+            return Arrays.asList(line.split(","));
+        } catch (FileNotFoundException e) {
+            throw new SQLSyntaxException("The table does not exist");
+        } catch (IOException e) {
+            System.err.println("There was an error reading the table file");
+        }
+
+        throw new SQLSyntaxException("The table does not match with a known format");
+    }
+
+    private static class Table {
+        private String tableName;
+        private List<String> columns;
+
+        public Table(String tableName, List<String> columns) {
+            this.tableName = tableName;
+            this.columns = columns;
+        }
+
+        public String getTableName() {
+            return tableName;
+        }
+
+        public void setTableName(String tableName) {
+            this.tableName = tableName;
+        }
+
+        public List<String> getColumns() {
+            return columns;
+        }
+
+        public void setColumns(List<String> columns) {
+            this.columns = columns;
+        }
     }
 }
